fix:  isAppConfigDifferent uses isEqual with Json.stringfy for comparison of objects
