+++ packages/examples/package.json
@@ -1,7 +1,7 @@
 {
   "name": "monaco-editor-wrapper-examples",
   "private": true,
-  "version": "3.4.0-next.3",
+  "version": "3.4.0-next.4",
   "type": "module",
   "dependencies": {
     "@codingame/monaco-vscode-configuration-service-override": "~1.83.5",
@@ -9,13 +9,13 @@
     "@codingame/monaco-vscode-json-default-extension": "~1.83.5",
     "@codingame/monaco-vscode-keybindings-service-override": "~1.83.5",
     "@codingame/monaco-vscode-python-default-extension": "~1.83.5",
-    "@typefox/monaco-editor-react": "~2.4.0-next.3",
+    "@typefox/monaco-editor-react": "~2.4.0-next.4",
     "http-server": "~14.1.1",
     "langium": "~2.1.1",
     "langium-statemachine-dsl": "~2.1.0",
     "monaco-editor": "npm:@codingame/monaco-editor-treemended@>=1.83.5 <1.84.0",
     "monaco-editor-workers": "~0.44.0",
-    "monaco-editor-wrapper": "~3.4.0-next.3",
+    "monaco-editor-wrapper": "~3.4.0-next.4",
     "monaco-languageclient": "~7.0.0",
     "monaco-languageclient-examples": "~7.0.0",
     "react": "~18.2.0",
