+++ packages/monaco-editor-wrapper/package.json
@@ -1,6 +1,6 @@
 {
   "name": "monaco-editor-wrapper",
-  "version": "3.4.0-next.3",
+  "version": "3.4.0-next.4",
   "license": "MIT",
   "description": "Monaco-Editor and Monaco Languageclient Wrapper",
   "keywords": [
