+++ packages/monaco-editor-wrapper/src/editorAppBase.ts
@@ -266,3 +266,11 @@ export const isModelUpdateRequired = (config: EditorAppConfigBase, modelUpdate: 
     const updateRequired = propsModelUpdate.some(propCompare);
     return updateRequired ? ModelUpdateType.MODEL : codeUpdate;
 };
+
+export const isEqual = (obj1: unknown, obj2: unknown) => {
+    if (obj1 instanceof Object && obj2 instanceof Object) {
+        return JSON.stringify(obj1) === JSON.stringify(obj2);
+    } else {
+        return obj1 === obj2;
+    }
+};
