+++ packages/monaco-editor-wrapper/test/editorAppClassic.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 import { EditorAppClassic, EditorAppConfigClassic } from 'monaco-editor-wrapper';
-import { createBaseConfig } from './helper.js';
+import { createBaseConfig, createEditorAppConfig } from './helper.js';
 
 const buildConfig = () => {
     const config = createBaseConfig('classic');
@@ -17,7 +17,7 @@ describe('Test EditorAppClassic', () => {
 
         const app = new EditorAppClassic('config defaults', config);
         expect(configclassic.$type).toEqual('classic');
-        expect(app.getConfig().editorOptions?.['semanticHighlighting.enabled']).toEqual(true);
+        expect(app.getConfig().editorOptions?.['semanticHighlighting.enabled']).toBeTruthy();
     });
 
     test('editorOptions: semanticHighlighting=false', () => {
@@ -26,7 +26,7 @@ describe('Test EditorAppClassic', () => {
         configclassic.editorOptions!['semanticHighlighting.enabled'] = false;
 
         const app = new EditorAppClassic('config defaults', config);
-        expect(app.getConfig().editorOptions?.['semanticHighlighting.enabled']).toEqual(false);
+        expect(app.getConfig().editorOptions?.['semanticHighlighting.enabled']).toBeFalsy();
     });
 
     test('editorOptions: semanticHighlighting="configuredByTheme"', () => {
@@ -36,5 +36,32 @@ describe('Test EditorAppClassic', () => {
 
         const app = new EditorAppClassic('config defaults', config);
         expect(app.getConfig().editorOptions?.['semanticHighlighting.enabled']).toEqual('configuredByTheme');
+    });
+
+    test('isAppConfigDifferent: basic', () => {
+        const orgConfig = createEditorAppConfig('classic') as EditorAppConfigClassic;
+        const config = createEditorAppConfig('classic') as EditorAppConfigClassic;
+        const app = new EditorAppClassic('test', createBaseConfig('classic'));
+        expect(app.isAppConfigDifferent(orgConfig, config, false)).toBeFalsy();
+
+        config.code = 'test';
+        expect(app.isAppConfigDifferent(orgConfig, config, false)).toBeFalsy();
+        expect(app.isAppConfigDifferent(orgConfig, config, true)).toBeTruthy();
+
+        config.code = '';
+        config.useDiffEditor = true;
+        expect(app.isAppConfigDifferent(orgConfig, config, false)).toBeTruthy();
+    });
+
+    test('isAppConfigDifferent: non-simple properties"', () => {
+        const config1 = buildConfig();
+        const config2 = buildConfig();
+        const configclassic1 = config1.wrapperConfig.editorAppConfig as EditorAppConfigClassic;
+        configclassic1.editorOptions!['semanticHighlighting.enabled'] = true;
+        const configclassic2 = config2.wrapperConfig.editorAppConfig as EditorAppConfigClassic;
+        configclassic2.editorOptions!['semanticHighlighting.enabled'] = true;
+
+        const app = new EditorAppClassic('config defaults', config1);
+        expect(app.isAppConfigDifferent(configclassic1, configclassic2, false)).toBeFalsy();
     });
 });
